Updated the validators and the converters to be more flexible
Removed isArrayOrSingle, the same functionality can be achieved using oneOf.
Changed the structure of the validators, now separate files for better overview.
Corrected the tests to work with the changes, more changes here is coming.
Added a new validator for notEmpty. This now needed to make sure something should not be empty, this was implicit before.
Now the converter can be generated from the validators as well as the default value.