Added a way to lazily load functions, to be used with actions & commands
Will make loading extensions much faster since we will not need to wait for a lot of require to complete.
This work will instead be delayed until the function is called meaning that the actual work of requiring files will only happen when we actually want to use something.
Using this in our extensions we saw a improvement in the loading of extensions up to 6 times faster in a example project using roc-package-web-app.