Updated how extensions are loaded and improved the state shape
Adds a way to see how long time the loading of the extensions took.
Improved the error message when a collision is detected for commands and settings.
Fixed a problem with the generated meta data for commands that would list the same extension multiple times.
Changed how init and postInit work along with changes to how the roc object is managed.
Fixed some small problems related to how extensions where loaded.
Will now take the description form package.json if one exists.