Changed command args and opts from arrays to objects, same for hook args
There are some reasons for this change.

Options are logical to be an object since the order never matters and maps well against the mental model of it.
Arguments are mentally structured as an array but by using an object we get some benefits. First of all we get the possibility to easily compose arguments in the same command together, this also holds true for options. This means that an extension can add extra arguments and options / modify existing ones without having to use init. Another reason is that we map the arguments in our commands to an object, making them fit well against then end output. This is not the case for arguments in hooks but by making the interface the same for all of them we make it easier to get into Roc with fewer different interfaces for similar things.

This change also makes it clear without knowing the implementation that name (now the key) is mandatory / highly recommended.