Updated template system to be more flexible and better overall
We do not any longer automatically install dependencies for the created project.
Changed how roc.setup.js(on) works to make it faster to create most projects, removing the need to install things in most cases.
Added support for full Handlebars syntax with the possibility to add additional helpers.
Made some work making it possible to extract the template system and the commands to a separate module in the future to keep the core more slim.
The current iteration is inspired by vue-cli.
Will add tests in the future.